requirements.txt
aiohttp>=3.9.0
networkx>=3.0
google-genai>=0.1.0
pydantic>=2.0.0
pytest>=7.0.0
